requirements.txt
GitPython==0.3.2.RC1
Twisted==13.1
blinker==1.3
pycrypto==2.6
glob2==0.4
PyYAML>=3